add transactions and core errors

+++ internal/service/user/service.go
@@ -4,19 +4,24 @@ import (
 	"context"
 	"errors"
 	"golang.org/x/crypto/bcrypt"
+	"municipality_app/internal/domain/core_errors"
 	"municipality_app/internal/domain/entity"
 	"municipality_app/internal/domain/repository"
 	"municipality_app/internal/domain/service"
 )
 
 func (svc userService) RegisterUser(ctx context.Context, data *service.CreateUserData) (*entity.User, error) {
+	if err := data.Validate(); err != nil {
+		return nil, err
+	}
+
 	userExists, err := svc.GetUserByEmail(ctx, data.Email)
 	if err != nil {
 		return nil, err
 	}
 
 	if userExists != nil {
-		return nil, errors.New("this email is already registered")
+		return nil, core_errors.EmailAlreadyUsedError
 	}
 
 	hashedPassword, err := hashPassword(data.Password)
@@ -32,18 +37,21 @@ func (svc userService) RegisterUser(ctx context.Context, data *service.CreateUse
 	}
 
 	user, err := svc.UserRepository.CreateUser(ctx, crateData)
-
 	return user, err
 }
 
 func (svc userService) UpdateUser(ctx context.Context, data *service.UpdateUserData) (*entity.User, error) {
+	if err := data.Validate(); err != nil {
+		return nil, err
+	}
+
 	userExists, err := svc.UserRepository.GetUserByID(ctx, data.ID)
 	if err != nil {
 		return nil, err
 	}
 
 	if userExists == nil {
-		return nil, errors.New("user no found")
+		return nil, core_errors.UserNotFound
 	}
 
 	if data.Name != nil {
@@ -61,7 +69,7 @@ func (svc userService) UpdateUser(ctx context.Context, data *service.UpdateUserD
 		}
 
 		if userFull != nil {
-			return nil, errors.New("this email is already registered")
+			return nil, core_errors.EmailAlreadyUsedError
 		}
 
 		userExists.Email = *data.Email
@@ -73,14 +81,22 @@ func (svc userService) UpdateUser(ctx context.Context, data *service.UpdateUserD
 }
 
 func (svc userService) Login(ctx context.Context, data *service.UserLoginData) (*entity.UserAuthToken, error) {
+	if err := data.Validate(); err != nil {
+		return nil, err
+	}
+
 	userFull, err := svc.UserRepository.GetUserFullByEmail(ctx, data.Email)
 	if err != nil {
-		return nil, err
+		return nil, core_errors.AuthErrorError
+	}
+
+	if userFull == nil || userFull.IsBlocked {
+		return nil, core_errors.AuthErrorError
 	}
 
 	compareErr := comparePassword(data.Password, userFull.Password)
 	if compareErr != nil {
-		return nil, compareErr
+		return nil, core_errors.AuthErrorError
 	}
 
 	err = svc.UserAuthService.DeleteExtraTokens(ctx, userFull.ID)
@@ -108,17 +124,19 @@ func (svc userService) GetUserByID(ctx context.Context, id int64) (*entity.User,
 	return svc.UserRepository.GetUserByID(ctx, id)
 }
 
-func (svc userService) BlockUserByID(ctx context.Context, id int) error {
-	//TODO implement me
-	panic("implement me")
+func (svc userService) BlockUserByID(ctx context.Context, id int64, isBlocked bool) error {
+	return svc.UserRepository.ChangeUserBlocked(ctx, id, isBlocked)
 }
 
-func (svc userService) ChangeUserIsAdmin(ctx context.Context, id int) error {
-	//TODO implement me
-	panic("implement me")
+func (svc userService) ChangeUserIsAdmin(ctx context.Context, id int64, isAdmin bool) error {
+	return svc.UserRepository.ChangeUserAdmin(ctx, id, isAdmin)
 }
 
 func (svc userService) ChangeUserPassword(ctx context.Context, data *service.ChangeUserPasswordData) error {
+	if err := data.Validate(); err != nil {
+		return err
+	}
+
 	hashedPassword, err := hashPassword(data.Password)
 	if err != nil {
 		return err
